Spring Boot 3.5, Gateway MVC, a plain Java core, and measured optimizations
- Three artifacts: gatepass-core (plain Java), gatepass-spring-boot-starter
  (Spring Boot 4) and gatepass-spring-boot3-starter (Spring Boot 3.5), the two
  starters compiled and tested from the same sources
- Spring Cloud Gateway Server MVC: signing, stripping and body signing, with
  load-balanced routes recognised through the load balancer
- The in-memory replay guard keeps 128-bit hashes in flat tables: 31 bytes per
  remembered pass instead of 120
- HMACs and digests are keyed once and cloned: issuing and verifying a pass is
  about a third faster
- Reactive bodies are read with one copy per chunk, in one place
- JMH benchmarks, shared test support, CI on both Spring Boot lines, a weekly
  job against the newest Spring releases, and Dependabot

+++ gatepass-spring-boot-starter/src/main/java/io/github/danimmll/gatepass/web/InboundRules.java
@@ -103,8 +103,8 @@ public final class InboundRules {
      *
      * @return the limit, in bytes
      */
-    public int maxBodyBytes() {
-        return (int) this.maxBodyBytes;
+    public long maxBodyBytes() {
+        return this.maxBodyBytes;
     }
 
     /**
